+++ lib/src/utils/response/http.dart
@@ -9,14 +9,14 @@ class ErrorModel implements Exception {
     this.reasonPhrase,
   });
   factory ErrorModel.fromJson(Map<String, dynamic> json) => ErrorModel(
-        errors: json['errors'],
-        type: json['type'] as String?,
-        reasonPhrase: json['reasonPhrase'] as String?,
-        title: json['title'] as String?,
-        status: (json['status'] ?? json['Status']) as int?,
-        detail: (json['detail'] ?? json['Detail']) as String?,
-        instance: json['instance'] as String?,
-      );
+    errors: json['errors'],
+    type: json['type'] as String?,
+    reasonPhrase: json['reasonPhrase'] as String?,
+    title: json['title'] as String?,
+    status: (json['status'] ?? json['Status']) as int?,
+    detail: (json['detail'] ?? json['Detail']) as String?,
+    instance: json['instance'] as String?,
+  );
   final dynamic errors;
   final String? reasonPhrase;
   final String? type;